requirements.txt
opencv-python~=4.11.0.86
screeninfo~=0.8.1
numpy
python-dotenv
google-generativeai
Pillow